+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.0.3
+
+* Fix initial value issue in AppDropDown
+
+## 0.0.2
+
+* AppDropDown now use data from InputDecorationTheme.of(context).
+
 ## 0.0.1
 
 * Initial Release with getDesignSize function and AppDropDown class.